requirements.txt
Django==5.1.3
requests==2.32.3
python-dotenv==1.0.1